Adding/adjusting tests
(1) Added a few tests to observe the behavior of cursors vs. greedy cursors when joining to replace blocks.

(2) Commented out the inverse-priority tests in testGrammarCorrespond and testGrammarReplaceEpsilon -- they're not just getting the wrong results, they cause evaluation to hang (because of the disagreement-about-finishing mentioned in the previous commit).

It's possible to test for this situation, like a mature greediness solution would simply say "Okay, given that this is my child, I can't be greedy", but for now I just want to see if this helps the efficiency problems.