Fixing SingleTape bug
When fixing the serious oversight Haokun found -- that programmers couldn't invoke a regex like `{VOWEL}n`, I introduced another bug.  It only affected erroneously-programmed grammars, so far as I can reckon, but even so it pointed to something being subtly incorrect in our semantics.

I had said in a recent commit that we had incorrectly left a check in calculateTapes -- the check that the children of a SingleTape only had one tape.  We can't do that check/fix until handleSingleTapes happens, or else it forbids reasonable regexes like `{VOWEL}n`.  So I got rid of that check and moved it later to handleSingleTapes.

But it wasn't an incorrect check after all!  It turns out handleSingleTapes is too late to fix that, because the fix can change the tape structure of the whole grammar, and handleSingleTapes can only fix up tape problems in the scope of a SingleTape.  This caused a few of our tests to fail (some erroneous grammars got fixed up in incorrect ways).

So we can't do handleSingleTapes after calculateTapes.  However, in a chicken-and-egg problem, we also couldn't do handleSingleTapes BEFORE calculateTapes, because handleSingleTapes needed to make a Rename, and to do that it needs to know what tape to rename from.  But Embeds don't have that information until calculateTapes.

But I also didn't want to combine them into a single pass -- calculateTapes is already complex enough.

So I rewrote how handleSingleTapes and calculateTapes divided responsibility for handling SingleTape+Embeds.

  (1) handleSingleTapes is only responsible for changing the scope of the SingleTape.  This whole sequence of bugs came down to the fact that SingleTapes scope over potentially complex grammars, but they don't actually mean "My child grammar only has one tape" (that would exclude {VOWEL}n), they mean "My child grammar is composed only of single-tape terminals."  So this pass eliminates the original SingleTapeGrammar, and "pushes" it further into the tree to the level of those terminals, so that it scopes tightly around the rather than around the whole sub-grammar.

    (1b) Actually, we just handle SingleTape(Literal) and SingleTape(Dot) right here, there's no reason to wait when they're literal tape names.  We really only put the new SingleTape around Embeds; that's the only one where we have to wait for future information.

  (2) The actual checking for single-tape status, fixing-if-erroneous, and tape renaming now happens in calculateTapes.  Unlike previously, though, this check is now happening at the appropriate scope.  Instead of checking the whole of `SingleTape({VOWEL}n)` and rejecting it, we're only checking SingleTape({VOWEL}).