Separating out Collection and Selection
Our old CollectionGrammar actually had a pretty complex lifecycle

    (a) starting out as a semantics-less container for assignments,
    (b) then being a flat structure (but retaining the logical structure of (a) for later name qualification)
    (c) same as (b) except with one symbol selected to be the semantics of the whole project

This commit separates out (a/b) and (c) into two different objects.  The former is still called CollectionGrammar, but the latter is now a SelectionGrammar.

They're still similar grammars in most respects and I'm not sure this refactoring was worth it, but it *did* force me to address a bit of sloppiness.  E.g. there was one situation remaining where we used a CollectionGrammar *without* specifying any symbol, and that's when running tests; now that a SelectionGrammar is the only valid thing to execute from, we have to choose an appropriate symbol (in this case, ALL).