Separating CollectionGrammars further
The old CollectionGrammar had a complex lifecycle:

    (a) Starts out as a potentially-nested hierarchy of itself
    (b) Then is flattened into a non-nested collection where all symbols have fully-qualified names
    (c) Finally, we select one symbol from it to be the "meaning" of the grammar as a whole.

The last commit split off (c) into SelectionGrammar; this commit splits off (a) and (b) into CollectionGrammar and QualifiedGrammar respectively.

Again this doesn't accomplish THAT much but it forces us to be specific about what this thing is and what's been done to it; if we're loosy-goosy about whether it's been flattened or selected we'll get type errors.  I think it helped me catch some tiny issues that might have been bugs or could become bugs later.

The trickiest bit here was reconciling this with how testHasVocab works, since that's patching a specific symbol within a collection.  For unclear reasons, splitting (a) and (b) caused several vocab tests to break in strange ways on nested collections -- I actually can't quite work out how these tests WEREN'T failing before, because it seems some of them would have been looking up fully-qualified names before flattening had occurred.

In any case, to get around this, we can flip the order of the symbol-patching and calling prepareInterpreter.  prepareInterpreter calls Interpreter.fromGrammar here, and that already has the mechanism to handle the three possible forms of grammars in testHasVocab (bare, Collection, and Dict<Grammar>) and also qualifies names as part of the process.  Then we can reach into the Interpreter and patch the relevant symbol.