Sorting tape names in interfaces
Now when you call Interface.getTapeNames(), it returns them in sorted order (ascending, case-insensitive).  And since both the Sheets add-on and the CLI call this function, they'll be in the same order in both places.

Also fixed a miniscule "bug": the highlight color of tapes depended on the capitalization of the tape name, which means that if the programmer is inconsistent about tape capitalization (which they're allowed to be, tape names are not case-sensitive), the same tape could receieve different colors in different places.  Now the highlight color depends on the lowercased tape name.