Fix to testHasVocab and default symbols
When the new implementation of testHasVocab was called with "" as its symbol, it was failing because it couldn't find any symbol by that name in interpreter.grammar.symbols.  (That symbol is actually named "Default".)  So this commit fixes that by replacing "" with "Default" in that situation.

For some reason these tests (and some others) weren't failing when I was running tests yesterday.  I'm thinking maybe I wasn't reloading -- that would cause THESE failed tests not to display in the interface, although it doesn't explain all the missing failed tests... In any case, now everything's passing.