Implemented a test suite for the Gramble CLI.
Also did much refactoring and bug fixes in the CLI itself. The refactoring
focussed around splitting out UI functionality and argument descriptions that
were common between the generate and sample commands, and also cleaning up the
error handling. The bug fixes were mostly based around verification of arguments.
Also fixed the max option for generate to actually limit the number of outputs.

There are still some more areas to test, in particular, the --format, --output,
and --verbose options for the generate and sample commands, but I wanted to
get a commit done for the bulk of the changes.