Added deterministic sampling using a seeded PRNG for testing purposes.
We use the seedrandom NPM library to provide the seeded PRNG.

seedrandom can be used to monkey patch Math.random, so no code changes are
needed in order to use a seeded PRNG. This is how we use seedrandom.

I modified testSamplingUtil.testSample()  so that all unit tests run with a seeded
PRNG. testSample now accepts a seed string, using a default one ('Seed-2024') if
one if not provided. testSample() now also accepts a results argument; if
provided, we skip calling generateOutputs() to produce the expected outputs.

I added a '--seed <string>' option to the sample command in the Gramble CLI.
If no --seed option is specified, we don't monkey patch to use a seeded PRNG,
instead using the normal Math.random. If the --seed option is specified, we use
a seeded PRNG; if no seed string is provided, the default 'Seed-2024' is used to
seed the PRNG.

I fixed a small bug in testUtil.testMatchOutputs(), where no error was generated
when no expected outputs were provided because no test was done.

I also fixed a small bug in testCLIUtil.testMatchOutputLines where empty outputs
were not handled correctly.

Finally, we added a few tests to testSampling to test that the seeded PRNG was
working as expected. We also changed testCLI to add the --seed option to all
sampling tests to make them deterministic, and also added a couple tests to test
that the seeded PRNG was working as expected.