Updated Parse tests for ^ and & being disallowed in symbols.
It turns out that some tests don't have the expected behaviour for ^, the
single-char-negation operator. This is because ^ (circumflex) is not a
punctuation character; it is actually in the Modifier Symbol (Sk) unicode
general category, so it is valid with symbol names because all unicode Modifier
Symbols are valid within symbol names.