Changed the single-char-negation operator from '^' to '!'.
The reason for this change is that '^' is in the Modifier Symbol (Sk) Unicode
class, which is permitted within symbols (including at the start) while '!' is
in the Punctuation class and is not valid within symbols.

Updated the Regular Expressions documentation page to reflect the change.
Also made some small corrections to the Regular Expressions and Characters
documentation pages.