Isolating replace/embed bug
I've isolated the bug that Chris found: that if he performed a replacement rule on a grammar, and then referenced it using embed in another symbol, the rule doesn't apply.  (I.e., you only get the base form of the word before replacement.)

I've discovered that it  only applies when you have another symbol embedded into the from/context of the rule.  This, combined with the fact that the rule *disappears*, rather than leading to an empty or ill-formed output, suggests that this is a problem of compilation, not execution.  A rule would only disappear like this if something was caught wrong with it.  We're not seeing an error, so that's suggesting the error is being fixed (by removal) but the error message is getting dropped.

Also, the thing that gets dropped is the entire rule block, not just the individual rule.  (E.g., if you have two rules and only one of them embeds a symbol, NEITHER apply, suggesting that the thing that's getting removed is the entire block.)