Narrowing down the embed/replace bug
I've further narrowed down the bug.  It's not about the ReplaceBlock being removed, the replaces are still present, the problem has to do with tokenization: the "Chris configuration" (an embed of a replace that itself has an embed) is causing the tape in question to resolve to an Atomic vocabulary, which is obviously going to screw up replacement.  That's why it looks like the replacement "isn't happening"; the tokens the replacement is looking for aren't there, they're still stuck in bigger morpheme-size chunks.

It seems like the error occurs during the vocab resolution phase of tape calculation, when we take the initial tape structures of each symbol in a QualifiedGrammar, then push that info back into the tree.  For a refresher: before this point, some of the vocab information is lazily unresolved -- the full answer would require information from other symbols, information that's not available until we make it available in this step.  In these cases, we couldn't yet work out the vocab, so we keepkept a tree representing the future calculation ahead.  As soon as the info about the other symbols becomes available, the tree collapses down into a real vocab object.

In the Chris configuration, this process gets the wrong answer: despite having correct information available, it's collapsing down into atomic literals.

I'm not sure why that is yet, but one hypothesis is that it has to do with the SingleTapes that get wrapped around embeds in the replacement rule: at this point, SingleTapes still think their tape is the dummy tape $T.  That's not correct and will be adjusted later in the process, but we struggled to think of any case where the lack of that information at this point in the algorithm would actually cause any changes to actual vocabs.  Perhaps this is one such case.