Isolating embed/singletape bug
I've further narrowed down the source of the bug; it's nothing to do with Replaces or ReplaceBlocks, just with SingleTapes.  We can duplicate the bug for Filters as well.  I'll concentrate on solving this for Filters since they're much simpler in structure.

Current hypothesis: grammars containing SimpleTape(Embed(name)) are prematurely collapsing into literals during the resolveAll stage.  It's probably due to the "false" tape structure that SingleTapes have at this stage: if their child doesn't have literal tapes, we temporarily treat it as a rename from the dummy tape $T.

Later, when we push this incorrectly-literal information back into the tree, references to those grammars have an incorrect tape structure.  (The grammar *itself* doesn't have an incorrect tape structure, that gets resolved correctly after the push.  Only grammars that reference it have the incorrect structure.)