Fixing embed/filter bug
I've fixed the filter case of the Embed/SingleTape bug.  This doesn't fix the replacement cases yet, but I think a similar strategy will.

There were two issues here:

1. Tape resolution for single tapes

When tapechecking SingleTapeGrammar, we were using a Tape.Rename as its representative in the tapeID structure.  This was the wrong thing to do, when there's an Embed in its scope.  As mentioned in the previous commit, what happens in this case is that we don't yet know what tape should be the toTape of the Tape.Rename, we default to the dummy tape $T.  The Tape.Rename erroneously looks for $T, can't find it, and defaults to an empty atomic vocab.

So using Tape.Rename isn't the right thing here; it's behavior is similar to what we want to happen in SingleTapes, but not the same.  So I introduced a new Tapes.Single that does what we need it to do: take the vocab of whatever it's child's tape is, and treat it as the vocab of its own tapeName.

2. Tape resolution for filters

There was also an issue with