Fixing embed/filter bug, take 2
(In the middle of writing the last commit message I saved and exited too soon.)

2. Tape resolution for filters

There was also an issue with tapecalc in filter conditions like StartsGrammar.  The way were were making the tape tokenized was to make an empty Tokenized vocab for each known tape and then sum it with the tapes of the child.  However, we couldn't do this when we didn't know what tapes the child had, like when there's an Embed inside, and that was leading to that tape not being tokenized.

But for an actual Gramble-language source file, we do know!  Syntactically, it's right there:

    starts text
    x

(That's the only possible syntax for "starts" now, by the way.  It used to be you could do "starts text" vs. "starts re text" vs. "starts embed", but now all three of those have been united into "starts text".)

We didn't actually store that "text" as part of the StartGrammar, only in its child, because of the possibilities mentioned above for "starts re text" and "starts embed".  Also, we allowed multi-tape filter conditions in testing, despite these being impossible to construct in syntax.  But I don't think we're ever going to enable multi-tape filter conditions in the actual Gramble syntax ever again -- our last big nailing-down-of-the-semantics, to fix Haokun's bug, wouldn't make sense if filters could be multi-tape.

So now, StartsHeader and its siblings actually store their tape name, and it's carried all the way to where StartsGrammar etc. are turned into more specific grammars.

This means we don't have to wait around to know what tape the Starts etc. are going to be; we already know it.

TODO: I should go back and do some refactoring after this, because fixed-tape filter conditions let us avoid some complexity in the code.  But I don't want to make any big refactorings until the Replace version of this bug is fixed, too.