Fixing embed/replace bug
Like the Filter version of this bug, the solution to the Replace version had two parts.

First, there was using Tapes.Simple, which we already did in the previous commits.

Secondly, there was making sure that the larger structure made it so tokenization happened appropriately.  This was trivially easy: it turned out that everything was in place.  (The Replace equivalent of the previous commit -- making Replaces have known tape names -- happened long ago.)  The only issue is that we were giving up on the calculation too early -- we were returning early in the case of non-Lit tapes.  But that's unnecessary and incorrect; there's nothing about the calculation that followed that required Lits.  Fixing this made everything work.