User now decides on path for deployments
We used to assume all deployment metadata would be stored in a subdirectory deployments/, a sister to packages/, docs/, etc., and that this directory would be .gitignored.  This raises a problem, though -- what if you want to track those deployments through git (e.g. in a separate private repo)?  Or back it up through some other syncing mechanism like Dropbox?

So instead, `npm run deploy` now asks for a path to the folder you want to use for deployment, rather than have it be relative to a particular magic directory.

The sheet will be named whatever the last folder in this path is, e.g. if the path is `../../../gramble_deploy/PatTest`, then the sheet will be called "Gramble (PatTest)".