Allowing deployment from main directory
We previously had to be in the packages/sheets_addon/ directory to run `npm run deploy`, but this was a little unweildy, and now that deployment locations are relative paths, it's annoying to have to ../.. out of that directory every time.

The snag: We have to run the deploy command with its workspace set to packages/sheets_addon, so when deploy.sh is invoked, $PWD is still `packages/sheets_addon`.  That ends up being disorienting -- you think it's going to be relative to the directory you're running `npm run deploy` from, but it ends up being relative to packages/sheets_addon anyway!

The solution: There's another variable, $INIT_CWD, that we can use to stay relative to the directory the user invoked the command in, not the workspace directory.