Deployment from absolute paths
This change lets us deploy from absolute paths, not just relative ones.

Usually this wouldn't be an issue, we'd just use whatever path the user typed, but as noted in the previous commit, we don't want to run these relative to $PWD, but from $INIT_CWD.  That means we had to add $INIT_CWD... but we were adding it to everything without checking if it was a relative path first!  Now we check.

There's a complication from Windows bash, which is even if you use a Unix-style path like `/C/Users/username/github`, it'll mangle that into `C:/Users/username/github`, so you can't use "/" as an indicator of absoluteness.  (We could invoke with "MSYS_NO_PATHCONV=1", that'll keep things as it was actually typed, but the resulting paths don't seem to work with all the commands we call.  So instead we just add another case to find paths that look like C:/, D:\, etc.)