Fixing empty-symbol bug
Darlene noticed that an empty cell under an embed was causing the page to appear to load forever.  It ended up being a simple fix but a hard one to find:

1. It wasn't actually loading forever, it was just throwing an exception and stopping, but leaving the "Loading..." text on the screen.

2. The exception was thrown when trying to display the tape-filter inputs in the sidebar, specifically when determining the appropriate colors for each tape.  We were trying to determine the color for "$T", but (by design) that's not a valid tape name.  Where was this coming from?

3. After much searching, it's from the miniParser in content.ts.  When you parse a content cell under most headers (e.g. when it's a string or regex), and that cell is empty, it generates the terminal CELL_EMPTY, which is expressed as a LiteralGrammar("$T", "").  For most headers, that's wrapped in a SingleTapeGrammar and so ends up becoming the appropriate LiteralGrammar for the appropriate tape.

However, we were ALSO re-using this CELL_EMPTY terminal when parsing an empty symbol.  And we didn't ever turn the resulting LiteralGrammar("$T", "") into anything else.  (EmbedGrammars are created when parsing the cell, not before or afterward.  This lets us do things like embed symbols in strings, alternate them, etc.)

So what we needed to do is:

    3.1. Don't re-use CELL_EMPTY here; make a special term SYMBOL_EMTPY that creates an EmbedGrammar with an empty string.

    3.2. Make sure the name resolution step ignores these (returns epsilon) rather than tries to resolve them.

(I guess we could do both in one fell swoop, just make SYMBOL_EMPTY create that epsilon, but I think I prefer having that empty embed exist for a while.  Part of what made this hard to track down was that I was expecting empty embeds to exist as some kind of EmbedGrammar("") object somewhere in the tree, and there never was.)

4. As a side fix, I realized that with the change to using rollup, we now have to build Gramble before deploying it.  (This used to happen in the browserify step, if I remember correctly.)  Since I kept forgetting to do this -- and thus deploying/testing stale versions -- I put a pre-deployment build into the "deploy" command.