Instructions for deploying to a pre-existing sheet
We figured out how to more easily deploy to a pre-existing Google Sheet (e.g. a newly created one, or one created by selecting File -> Make a copy).  It turns out that `clasp clone` does the finicky work of making the metadata for you, meaning that all you really have to do is make a local directory, clone, and then deploy like usual.

Explaining this requires explaining a little more about what clasp is and does, so I added that to the README as well.