Tests to show scope of escaped space bug
We noticed recently that an escaped space literal at the end of a string doesn't survive; it becomes a literal backslash.  (Note that Gramble, unlike most programming languages, doesn't have a quotes around literal strings, so there's an additional dilemma on how to express "foo" vs "foo " that other languages don't have.)

In gramble the latter is expressed

    foo\

on its own in a cell; note that there's a space after it.  The problem is that whitespace gets trimmed early, then there's no space there to escape.