Fixing escaped space bug
As mentioned, we had a problem with cells like "foo\ "; since cells are trimmed during initial parsing, this space disappears and the resulting cell contains "foo\" rather than "foo".

There are two routes to fixing this:

(1) Don't trim early.  The cell parsers know what to do with whitespace-vs-escaped-whitespace, so let them handle it.

(2) Add the space back in if the trimmed cell ends with "\".  This sounds like a hack but it has the nice property that it makes "foo\ " and "foo\" behave identically -- these two cells LOOK identical but aren't, so it would be surprising to the programmer and difficult to debug if they lead to different results.

I ended up doing a bit of both.  We don't trim early any more, and pass the original untrimmed string to the cell parsers; in the handful of places where trimming is needed before this we just do so locally.

But I also add a space during parsing if a cell ends with "\" anyway, because I like the consequence of (2).

This caused/revealed a few remaining bugs:

(a) I haven't yet fixed how this affects rename/hide cells.  That's the sole major "unparsed" cell type remaining (i.e. onces that we parse by string wrangling rather than our parser).  We should probably use a parser for those.

(b) There's a bug with escaped backlashes still escaping what comes after them.  I discovered this while testing these changes, but it must have already existed; it occurs in contexts that would not have been affected by these changes.