Fix for double backslash problem
This addresses the second bug revealed in the previous commit.

Looking online for how to adapt our regexes to handle double backslashes correctly, several comments suggested that regexes were a poor fit anyway for escaping, that it was very difficult to get them to handle corner cases, and that the imperative solution ends up more reliable.  So now we have a simple loop through the string, looking for backslashes, and handling them as is contextually appropriate.