Parsing hide/rename cells
As I mentioned two commits ago, we had never actually been parsing the content cells under Hides and Renames, we had just been doing ad-hoc string parsing.  This meant that there were various kinds of errors we couldn't catch and potentially some unanticipated brittleness.  (I can't *quite* think of a case where it would have led to seriously erroneous output, since the errors I can think of would be caught in tape-checking, but now we can give better and more specific error messages at least.)

Now, there's a "tapename" ParseClass and TAPENAME_ grammars in content.ts, as well as a TapeNameGrammar to hold the results of these parses.  (It's not exactly a Grammar in the sense that it doesn't correspond to any Language in the formal sense, but at that stage in compilation stuff gets called Grammars, so this is a Grammar too.)

I've added a lot of tests for Hide/Rename edge cases that we hadn't tested for before.