Fix to parse tests
A small change in the previous commit broke some tests.  Specifically, I made it so erroneous symbol parses resulted in EmbedGrammar("") rather than EpsilonGrammar().  This broke a few tests that expected an epsilon and got an empty embed.

Trying to make erroneous symbol parses ALWAYS be an empty embed, though, proved a fool's errand.  It's fine for testParseSymbol, but it becomes difficult to predict whether a given erroneous regex parse in testParseRegex should be an epsilon or empty embed.  (It *is* predictable... just frustrating to do so.)

So I went back on that change in the previous commit: erroneous symbol parses result in epsilons again, and the tests pass.