Tracking down filter slowness
I've made a minimal grammar for a made-up language but with a similar overall structure to Yanfei's grammar.  This has revealed why some grammars end up particularly slow.

The problem isn't a bug, it's inherent to the problem of choosing what tapes to execute in what order.

The structure of this grammar is that there's a big table of roots, where each root has a class, and when we want to use it we filter it with an "equals class: whatever".  (These happen to be separated out into separate symbols for convenience, but that's not important for the slowdown.)

The issue is that we can get into situations where, whether we choose to execute the "text" tape first or "class", we go down garden paths.  Say the text is "faba" in class F, and there are lots of classes A, B, C, etc.

Say you're trying to parse/sample and you're given "text:faba", and start with the "text" tape. The engine has to go down one of the paths where we filter the class, but we don't know which one is going to be appropriate yet, we haven't done anything about the other tape.  Say we go down the path that filters the Root table with "class:B".  This will succeed on text -- "faba" is in the Root table after all.  It has "class:F", of course, but we haven't got there yet.  Having finished finding "faba", we now move on to the "class" tape, and immediately fail because we try to join "B" and "F".

On the other hand, if we do "class" first, this also leads us down garden paths.  We have to choose a class at the outset, so we don't know what to choose.  We might choose "B", then go through the whole process of finding "faba" and fail.  (Granted, this simple example fails quickly, but imagine if there were multiple filters.  We might be going through a whole tree of possible filter choices before hitting the one that correctly lets us parse "faba" in "text".

Anyway, addressing this will be a fairly large change, since it strikes at the idea that we might not be able to choose one best order of tape execution.  Rather, it'll probably be necessary to add some "pre-filtering" -- that is, even if we choose to do "text" first, we have a pre-compiled or just-in-time prefiltered "class:B" version of the Root table.

(That's probably unconsciously what we as Gramble programmers think happens anyway.  You put a filter on an embedded table and think "Okay, that's only going to be the lines that fulfill that filter" as if it's already happened.  But it's not, that filter might be the LAST thing the engine executes.  So this proposed change is basically making it so the engine works more like programmers might think it does.)