New operand tests, error testing improvements, better error messages.
Started writing operand tests (for replacement rules). More needs to done on this.

I discovered that we were not reporting missing operands when no operands
were specified. This required a small fix to CheckNamedParams.checkRequiredParams
to check even when there are no child headers.

While trying to sort out the generated errors, I decided to enhance our error
testing by allowing the programmer to optionally specify the expected short
message in the test, and we now verify that short message in the output. I might
consider adding optional long error message checking too in the future. Also, we
are not testing warning messages currently because we don't currently have
meaningful short messages for warnings. We should add short messages for warnings,
then we could test them too.

I added short error messages to the expected errors in the tests in testSourceError
and testSourceOperand.

In addition, I modified most generated short messages to include the cell/header
text in order to make the short error messages more meaningful.

Finally, I changed the contents of the CSV files for a few tests for
testSourceError to correct instances where some cells intended to be assignments
had not had ":" changed to "=".