Allowing from/to/context as identifiers
From/to/context were reserved words and couldn't be used as identifiers, but "context" especially is a pretty reasonable thing to call something.

The reason they were reserved words is because of their special role in naming parameter columns for the "replace" operator.  However, it's a bit of a surprise for parameter names to be reserved words: in other languages, just because a parameter is named "x" doesn't mean you can't name something "x" elsewhere.

So this relaxes the parsing so that "from"/"to"/"context" are allowed as identifiers, and thus parsed as ordinary TapeHeaders.  Later, we check whether they're in the right position to be the parameters of Repalce, and if so transform them to FromHeader etc.

This doesn't affect the "unique" param; "unique" is still, well, unique.  It needs a special parsing rule to allow us to parse strings like "unique text"; this syntactically can't be a tape name.  (That's also the old reason why from/to/context were special; you could say things like "from underlying" and "to surface".  We got rid of that construction, however, because in the general case we were having trouble compiling it into something that could run efficiently.)

This commit only gets things working, allowing from/to/context to be ordinary tapes in ordinary situations while still having them work specially in Replace; I haven't yet looked at error messaging or aligned the error messages with the new sourceOperand tests.