Fixed operand error testing issues revealed by the from/to/context-as-identifiers change.
The expected results for testHeaderID for "from", "to", "context" needed updating.

Secondly, the error messages for the tests in testSourceOperand changed. As a
consequence of the from/to/context-as-identifier changes, we started getting the
"This op has erroneous parameters and will not execute" warning messages for many
tests that previously did not generate that warning. I'm not entirely sure why we
didn't get this warning previously, but I think it may be that our "Invalid header"
errors were detected in a different spot. So I had to add those warnings where
appropriate in the expected errors for the tests.

Furthermore, it turns out that when we issue the "erroneous parameters" warning,
testing that the required "from" and "to" operands are present for the "replace"
operator was also not being done. I changed the order of checking things in
CheckNamedParams.handleOp, so these errors are now caught and reported too.