Inconsistent capitalization led to double loading
We found a bug that initially seemed to require a very complicated setup involving collections, multiple sheets, and a reference to a symbol on that sheet AND to the 'All' symbol in the collection, but where the two references were using different capitalization.  Turns out there are two bugs interacting with each other.  Each one was pretty simple once isolated.

(1) An unneccessary source-load when asking for an ALL symbol.

During normal compilation, we go through all the passes in order.  During source loading, we only go through the source passes, and then try to resolve names.  Unresolveable names are added to a queue and we try to load them as source.  (E.g. if we have Foo.Verbs we look for a sheet named 'Foo' or a file called 'Foo.csv'.)

But we missed a step!  In between source passes and name resolution we have a pass that creates symbols like ALL.  We forgot to run that during source passes.  So when you reference Foo.ALL, you initially get a missing name error and we look for it in Foo/Foo.csv.  But we never actually see that cause an error message: it's not an error to look for Foo/Foo.csv and not find it, and it's also not a missing name error in the end (Because by the end we DO perform that pass).  So it was an invisible error.

(2) A redundant source-load when external references used different capitalization (for example, asking for verbs.intransitive and Verbs.transitive).

Although symbols were case insensitive, we didn't take that into account when loading sources for missing names.  So if you referenced 'verbs' and 'Verbs' and neither was found, BOTH would be queued up for loading.  Then we'd end up loading that document twice, and then during compilation say 'Hey wait there are two symbols with identical names' (since these ARE the same name).

Anyway, both are fixed now.