Options and logging improvements
I made options a property of the DevEnvironment interface, not the Interpreter,
just like they are a property of Env. I also added a logDebug method to both
DevEnvironment and Env to make calling logDebug easier.

I changed the Options function so that it first removes entries with undefined
values before merging the provided options with the default options. This means
that undefined values will be replaced with their defaults. This also means that
the various testXxx<Aux> functions do not need to initialize their various
options parameters to their defaults -- this is now left up to Options.

I added a VERBOSE_EXPR verbose option along with a logExpr method in DerivEnv to
use for logging expressions instead of logDebug. This way logDebug can be used
for specific debugging situations in specific tests without getting a dump of
the grammar expressions if the expression dump is not wanted. That is, logDebug
is for less permanent logging situations. (Note: logDebugId and logDebugOutput
in DerivEnv become logExprId and logExprOutput.)

The logDebug function now accepts multiple msg arguments of any type as multiple
arguments for console.log, i.e. multiple fields on the same line, instead of
multiple lines of output.

I removed the verbose function from testUtil because its usage was very limited
-- I think it's better to call logDebug directly.